Fixes of the custom step settings
- Switched scoreStep from float to double
- Rename from BottomSheetPreference to ScoreStepsPreferenceSheet

+++ core/ui/src/main/java/com/axiel7/anihyou/core/ui/composables/scores/RatingView.kt
@@ -48,7 +48,7 @@ fun RatingView(
     rating: Double? = null,
     showAsDecimal: Boolean = false,
     decimalLength: Int = if (showAsDecimal) 1 else 0,
-    increments: Float = 1.0f,
+    increments: Double = 1.0,
     onRatingChanged: (Double?) -> Unit,
 ) {
     val haptic = LocalHapticFeedback.current
