+++ feature/settings/src/main/java/com/axiel7/anihyou/feature/settings/SettingsEvent.kt
@@ -61,7 +61,8 @@ interface SettingsEvent : UiEvent {
     fun setStaffNameLanguage(value: UserStaffNameLanguage)
 
     fun setScoreFormat(value: ScoreFormat)
-    fun setScoreStep(value: Float)
+
+    fun setScoreStep(value: Double)
 
     fun setDefaultTab(value: DefaultTab)
